+++ testsuite/tests/emcy_hist.c
@@ -37,6 +37,7 @@ TS_DEF_MAIN(TS_Emcy_HistNumOnSet)
     uint8_t     hist_num;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -59,6 +60,7 @@ TS_DEF_MAIN(TS_Emcy_HistNumOnSetRepeat)
     uint8_t     hist_num;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -82,6 +84,7 @@ TS_DEF_MAIN(TS_Emcy_HistNumOnSetMulti)
     uint8_t     hist_num;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -105,6 +108,7 @@ TS_DEF_MAIN(TS_Emcy_HistNumOnClr)
     uint8_t     hist_num;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -130,6 +134,7 @@ TS_DEF_MAIN(TS_Emcy_HistNumOnReset)
     uint8_t     hist_num;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -155,6 +160,7 @@ TS_DEF_MAIN(TS_Emcy_HistObjOnSet)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -177,6 +183,7 @@ TS_DEF_MAIN(TS_Emcy_HistObjOnSetRepeat)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -200,6 +207,7 @@ TS_DEF_MAIN(TS_Emcy_HistObjOnSetMulti)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -226,6 +234,7 @@ TS_DEF_MAIN(TS_Emcy_HistObjOnClr)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -254,6 +263,7 @@ TS_DEF_MAIN(TS_Emcy_HistObjOnReset)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1                        */
@@ -282,6 +292,7 @@ TS_DEF_MAIN(TS_Emcy_HistClearOk)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1 without user info      */
@@ -315,6 +326,7 @@ TS_DEF_MAIN(TS_Emcy_HistClearBad)
     uint32_t     hist;
                                                       /*------------------------------------------*/
     TS_CreateMandatoryDir();
+    TS_CreateEmcy();
     TS_CreateNode(&node,0);
 
     COEmcySet(&node.Emcy, 1, 0);                      /* register error #1 without user info      */