+++ CHANGELOG.md
@@ -7,8 +7,18 @@ and starting with version 4.1.0 this project adheres to [Semantic Versioning](ht
 
 ## [unreleased]
 
+nothing
+
+## [4.1.5] - 2021-02-11
+
+### Added
+
+- Allow disabling the EMCY module by setting NULL for emcy code table.
+
 ### Fixed
 
+- Explain how to setup EMCY in quickstart example (even if not used in this application).
+- Fix NMT reset when EMCY is not used (no EMCY code table).
 - Fix broken RPDO communication setup [cjardin112/master](https://github.com/cjardin112).
 
 ## [4.1.4] - 2020-12-15
@@ -103,7 +113,8 @@ and starting with version 4.1.0 this project adheres to [Semantic Versioning](ht
 - First Open Source Release.
 
 
-[unreleased]: https://github.com/embedded-office/canopen-stack/compare/v4.1.4...HEAD
+[unreleased]: https://github.com/embedded-office/canopen-stack/compare/v4.1.5...HEAD
+[4.1.5]: https://github.com/embedded-office/canopen-stack/compare/v4.1.4...v4.1.5
 [4.1.4]: https://github.com/embedded-office/canopen-stack/compare/v4.1.3...v4.1.4
 [4.1.3]: https://github.com/embedded-office/canopen-stack/compare/v4.1.2...v4.1.3
 [4.1.2]: https://github.com/embedded-office/canopen-stack/compare/v4.1.1...v4.1.2